+++ packages/li-editor/src/widgets/LayersPanel/LayerAttribute/LayerForm/StyleForm.tsx
@@ -3,8 +3,8 @@ import {
   ColorRangeSelector,
   FieldSelect,
   FormCollapse,
-  IconList,
   IconSelector,
+  IconScaleSelector,
   Offset,
   ScaleSelector,
   ResterScaleSelector,
@@ -59,8 +59,8 @@ const StyleForm: React.FC<StyleFormProps> = (props) => {
           SliderRange,
           ScaleSelector,
           ResterScaleSelector,
-          IconList,
           IconSelector,
+          IconScaleSelector,
           ...registerForm.components,
         },
       }),
