+++ packages/li-p2/src/LayerAttribute/IconImageLayerStyle/constant.ts
@@ -2,14 +2,14 @@ import type { IconImageLayerStyleAttributeValue } from './types';
 
 /** 组件名称, 前缀 */
 export const CLS_PREFIX = 'li-p2-icon-image-layer-style-attribute';
-import { DEFAULTICONOPTIONS } from '../components/IconSelector/constant';
+import { DEFAULT_ICON_CATEGORY } from '../components/IconScaleSelector/constant';
 
-export const BuiltInImageList: { id: string; image: string }[] = DEFAULTICONOPTIONS.map((item) => item.icons).flat();
+export const BuiltInImageList: { id: string; url: string }[] = DEFAULT_ICON_CATEGORY.map((item) => item.icons).flat();
 
 export const BuiltInImage = BuiltInImageList.reduce(
-  (pre, { id, image }) => ({
+  (pre, { id, url }) => ({
     ...pre,
-    [id]: image,
+    [id]: url,
   }),
   {},
 );
