+++ packages/li-p2/src/LayerAttribute/components/IconScaleSelector/CustomItem/IconPanel/style.ts
