+++ packages/li-p2/src/LayerAttribute/components/IconScaleSelector/UnknownIcon/index.tsx
