+++ packages/li-p2/src/LayerAttribute/components/IconScaleSelector/constant.ts
@@ -1,0 +1,208 @@
+import type { IconOptionsType } from './type';
+
+export const DEFAULT_ICON_CATEGORY: IconOptionsType = [
+  {
+    type: '地标',
+    icons: [
+      {
+        id: 'landmark_park',
+        name: '公园',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*JZf9T6psSzkAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'landmark_fire-station',
+        name: '消防站',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*lKIITZAG_PcAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'landmark-coffee',
+        name: '咖啡馆',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*uHYERpUeZ_MAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'landmark_supermarket',
+        name: '商场',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*p2rgTYaHGdkAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'landmark_zoo',
+        name: '动物园',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*ElxdRIYcsS0AAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'landmark_railway-station',
+        name: '火车站',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*5Wg1TIbcCcYAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'landmark-mall',
+        name: '商城',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*jjynQJDfKigAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'landmark_nosocomium',
+        name: '医院',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*weuqSamNRnUAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'landmark_library',
+        name: '图书馆',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*XB8vR7azNvgAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'landmark_bus-stop',
+        name: '公交车站',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*LrCnSr8FeToAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'landmark_cinemas',
+        name: '电影院',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*OGFDT4Vyoz4AAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'landmark_dwelling',
+        name: '小区',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*oxcETa4MKgwAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'landmark_bank',
+        name: '银行',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*sWHeRrJvTesAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'landmark_cinema',
+        name: '剧院',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*uCdPTZrfSqIAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'landmark_amusement-park',
+        name: '游乐园',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*SIU-R7jDaLAAAAAAAAAAAAAADmJ7AQ/original',
+      },
+    ],
+  },
+  {
+    type: '天气',
+    icons: [
+      {
+        id: 'weather_sunny',
+        name: '晴天',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*polsQa3R6UgAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'weather_cloudy',
+        name: '多云',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*xOD4SIwHAMcAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'weather_cloudy-day',
+        name: '阴天',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*wMJISZvc_1EAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'weather_sand',
+        name: '扬沙',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*4nNTRYMV-iwAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'weather_tornado',
+        name: '龙卷风',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*ldyiQpgiJlYAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'weather_rainstorm',
+        name: '雷雨',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*E5JkRLtSzZYAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'weather_spit',
+        name: '小雨',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*nxwXSo_PR2sAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'weather_moderate-rain',
+        name: '中雨',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*4yRYRI3AwowAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'weather_heavy-rain',
+        name: '大雨',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*UeChTbbwvloAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'weather_sleet',
+        name: '雨夹雪',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*zi53Rp5hWMIAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'weather_light-snow',
+        name: '小雪',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*F-w5Qas4Np4AAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'weather_moderate-snow',
+        name: '中雪',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*GHisS4y9ynMAAAAAAAAAAAAADmJ7AQ/original',
+      },
+    ],
+  },
+  {
+    type: '医疗',
+    icons: [],
+  },
+  {
+    type: '消防',
+    icons: [
+      {
+        id: 'fire_fighting-car',
+        name: '消防车',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*zUjqQYt67SgAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'fire_fighting-tied',
+        name: '消防拴',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*iI5fSpCQWp0AAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'fire_fighting-extinguisher',
+        name: '灭火器',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*vIadR56-pKMAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'fire_fighting-water-pump',
+        name: '消防水泵',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*TFzHSLnqc9gAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'fire_fighting-pump-room',
+        name: '消防泵房',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*bWYLQqc8sTIAAAAAAAAAAAAADmJ7AQ/original',
+      },
+    ],
+  },
+  {
+    type: '其他',
+    icons: [
+      {
+        id: 'other_aquatic-germplasm-reserve',
+        name: '水产种质资源保护区',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*sSbCRKP1RsgAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'other_marine-nature-reserves',
+        name: '海洋自然保护区',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*dPoYQoBq4f0AAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'other_marine-park',
+        name: '海洋公园',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*on7pR5VRjHsAAAAAAAAAAAAADmJ7AQ/original',
+      },
+      {
+        id: 'other_special-marinep-protected-areas',
+        name: '特别海洋保护区',
+        url: 'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*DulySrR5zuwAAAAAAAAAAAAADmJ7AQ/original',
+      },
+    ],
+  },
+];
