+++ packages/li-p2/src/LayerAttribute/components/IconScaleSelector/index.md
@@ -9,7 +9,7 @@ nav:
   path: /components
 ---
 
-# icon 列表
+# icon 自定义
 
 ## 介绍
