+++ packages/li-p2/src/LayerAttribute/components/IconSelector/index.md
@@ -9,7 +9,7 @@ nav:
   path: /components
 ---
 
-# icon 自定义
+# icon 列表
 
 ## 介绍
