+++ packages/li-p2/src/LayerAttribute/components/index.ts
@@ -7,7 +7,7 @@ export {
   Slider,
   SliderRange,
 } from '../../components/Formily';
+export { default as IconScaleSelector } from './IconScaleSelector';
 export { default as IconSelector } from './IconSelector';
-export { default as IconList } from './IconList';
 export { default as ResterScaleSelector } from './ResterScaleSelector';
 export { default as ScaleSelector } from './ScaleSelector';
