+++ packages/li-p2/src/LayerAttribute/index.ts
@@ -3,7 +3,7 @@
 /**
  * 自定义 formily 图层组件
  */
-export { IconList, IconSelector, ScaleSelector, ResterScaleSelector } from './components';
+export { IconSelector, IconScaleSelector, ScaleSelector, ResterScaleSelector } from './components';
 
 /**
  * 图层属性面板
@@ -36,7 +36,7 @@ export { hexbinLayerStyleConfigToFlat, hexbinLayerStyleFlatToConfig } from './He
 export type { HexbinLayerStyleAttributeValue } from './HexbinLayerStyle/types';
 
 export { IconImageLayerStyleAttribute, IconImageLayerStyleAttributeSchemaField } from './IconImageLayerStyle';
-export { DefaultIconImageLayerStyle, BuiltInImage as DefaultIconList } from './IconImageLayerStyle/constant';
+export { DefaultIconImageLayerStyle } from './IconImageLayerStyle/constant';
 export { iconImageLayerStyleConfigToFlat, iconImageLayerStyleFlatToConfig } from './IconImageLayerStyle/helper';
 export type { IconImageLayerStyleAttributeValue } from './IconImageLayerStyle/types';
